"Add JWT-based user identifier for rate limiting purposes
Rate limiting policy has been modified to use JWT token-based user identifier instead of the IP address, preventing token sharing for bypassing rate limit. If the JWT token is invalid or expired, it falls back to the IP method.

The `setErrorHandler` function was also updated to deliver cleaner error messages by nesting the `message` inside an `error` object. This fosters better communication of error status with the client-side.

Finally, loading of routes and middlewares was refined for better efficiency, and JSDoc was added for better documentation